Change serializers to be commands
